move nodemap to lib
